In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,14 +6,14 @@
  */
 int main(void)
 {
-	 unsigned long int n_1 = 0, n_2 = 1, new, c;
+	 unsigned long long int n_1 = 0, n_2 = 1, new, c;
 
 	for (c = 0; c <= 97; c++)
 	{
 		new = n_1 + n_2;
 		n_1 = n_2;
 		n_2 = new;
-		printf("%lu", new);
+		printf("%llu", new);
 		if (c != 97)
 		{
 			printf(", ");
